Split off AreaOfInterest.
